% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_file_builders.R
\name{add_image_footnote}
\alias{add_image_footnote}
\title{Add Image Footnote}
\usage{
add_image_footnote(img_tag, footnotes, mode = "number")
}
\arguments{
\item{img_tag}{image tag that needs the footnote}

\item{footnotes}{the vector of the footnotes to add}

\item{mode}{number or alphabet mode for the superscript}
}
\value{
html string of characters
}
\description{
This function adds footnotes to images
}
\examples{
\dontrun{
add_image_footnote()
}
}
