% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymizer.R
\name{anonymize_redcap_data}
\alias{anonymize_redcap_data}
\title{Anonymize Redcap Data}
\usage{
anonymize_redcap_data(
  date_shift = TRUE,
  date_shift_each_id = TRUE,
  hash_ids = TRUE,
  hash_facility = TRUE,
  anonimize_ids = FALSE,
  anonimize_facility = FALSE,
  hash_key = "fullset",
  df = NULL
)
}
\arguments{
\item{data}{Tibble}
}
\value{
nothing
}
\description{
Helps address phi concerns when distributing redcap data
}
\examples{
\dontrun{
anonymize_redcap_data()
}
}
