% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatic_column_processor.R
\name{automatic_constructor}
\alias{automatic_constructor}
\title{Automatic Constructor}
\usage{
automatic_constructor(
  fields,
  event = "baseline_arm_1",
  instrument = NA,
  instance = NA,
  recode = list(),
  missing_val = NA,
  combine_checkbox = TRUE,
  mode = "data",
  recode_before_combine = FALSE
)
}
\arguments{
\item{fields}{a vector of fields to add to the analytic set}

\item{event}{a vector of length 1}

\item{instrument}{a vector of length 1}

\item{instance}{a vector of length 1}

\item{recode}{a list of recodings of all fields or a list of lists with length == fields length of recodings of each field}

\item{missing_val}{replace all na with this value}

\item{combine_checkbox}{unites checkboxes before recoding}

\item{mode}{either "data", "code", "columns", "count", or "fields"}

\item{recode_before_combine}{this will recode a checkbox before merging it with its other checkbox fields}
}
\value{
joined columns
}
\description{
Helps automatically create simple redcap columns in the analytic set
}
\examples{
\dontrun{
automatic_constructor()
}
}
