% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_tools.R
\name{boolean_column_counter}
\alias{boolean_column_counter}
\title{Boolean Column Counter}
\usage{
boolean_column_counter(df, groups = NULL)
}
\arguments{
\item{df}{tibble with boolean columns to count}

\item{groups}{optional grouping variable}
}
\value{
summed tibble
}
\description{
counts the boolean columns of a tibble
}
\examples{
\dontrun{
boolean_column_counter()
}
}
