% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_explorer.R
\name{bootstrap_new_study}
\alias{bootstrap_new_study}
\title{Bootstrap New Study}
\usage{
bootstrap_new_study()
}
\value{
nothing
}
\description{
use this to bootstrap a new study
}
\examples{
\dontrun{
bootstrap_new_study()
}
}
