% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytic_builder.R
\name{build_analytic_dataset}
\alias{build_analytic_dataset}
\title{Build Analytic Data Set}
\usage{
build_analytic_dataset(names = NULL, error_path = NULL)
}
\arguments{
\item{names}{you can specify a sublist of columns if you want}

\item{error_path}{you can specify a path to write the error dataframe to a csv}
}
\value{
analytic_dataset dataframe
}
\description{
integrates only test and error free contruct columns into full dataset
}
\examples{
\dontrun{
require_study()
}
}
