% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_explorer.R
\name{check_all_constructs}
\alias{check_all_constructs}
\title{Check all constructs}
\usage{
check_all_constructs(studies = NULL)
}
\arguments{
\item{studies}{if list of studies is passed than check will be perfect otherwise its aproximate}
}
\value{
vector of functions that do not comply with construct function naming standards
}
\description{
returns vector of all inappropriately named construct functions in this package
}
\examples{
\dontrun{
check_all_constructs()
}
}
