% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_writer.R
\name{check_order_functions}
\alias{check_order_functions}
\title{Check Order Functions}
\usage{
check_order_functions(names)
}
\value{
vector of function names ordered
}
\description{
orders the names of a study
}
\examples{
\dontrun{
order_functions()
}
}
