% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals_manager.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear Cache}
\usage{
clear_cache()
}
\value{
nothing
}
\description{
clears the cache of already processed get_construct calls
}
\examples{
\dontrun{
clear_cache()
}
}
