% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_tools.R
\name{column_unzipper}
\alias{column_unzipper}
\title{Column unzipper}
\usage{
column_unzipper(df, name, sep = ", ")
}
\arguments{
\item{df}{data to unzip the columns}

\item{name}{of column to separate}

\item{sep}{string to separate columns, default is ','}
}
\value{
unzipped tibble
}
\description{
takes a dataframe with comma separated list and widens it
into multiple columns, with boolean values
}
\examples{
\dontrun{
column_unzipper()
}
}
