% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_tools.R
\name{confirm_stability_of_related_visual}
\alias{confirm_stability_of_related_visual}
\title{Confirm Stability of Related Visual}
\usage{
confirm_stability_of_related_visual(function_name, key)
}
\arguments{
\item{function_name}{name of the visualization function}

\item{key}{verification key}
}
\value{
summed tibble
}
\description{
Makes sure a related visualization has not changed
}
\examples{
\dontrun{
confirm_stability_of_related_visual()
}
}
