% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_tools.R
\name{count_split_cols_long_and_wide}
\alias{count_split_cols_long_and_wide}
\title{Count Split Cols Long and Wide}
\usage{
count_split_cols_long_and_wide(df, seperator, remove_na = TRUE)
}
\arguments{
\item{df}{analytic dataset}

\item{seperator}{string to separate columns}

\item{remove_na}{whether nas ought to be removed}
}
\value{
dataframe
}
\description{
splits columns with a given seperator, then counts the values betwwen
the resultant groupt
}
\examples{
\dontrun{
count_split_cols_long_and_wide()
}
}
