% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_data_handler.R
\name{download_custom_data}
\alias{download_custom_data}
\title{Download Custom Data}
\usage{
download_custom_data(study, token, path = NULL, type = "csv", date = NULL)
}
\arguments{
\item{study}{The study name}

\item{token}{The redcap api token}

\item{path}{This is a default parameter that will auto name if given a directory 
path or will save to an arbitrary file if given a full path or will save to working directory}

\item{type}{This is a default parameter that will specify the type of file to download
it defaults to 'csv' but you can also request: a preparsed .rds with 'rds' an all character .rds with 'cds'
or a preparsed .feather with 'reg' an all character .feather with 'chr'}

\item{date}{This is a default parameter that you don't need to specify,
but if you do it will try to collect data from that date if its available.}
}
\value{
dataframe of Redcap Data
}
\description{
You can use this function to get arbitrary data from METRC studies
}
\examples{
\dontrun{
download_custom_data()
}
}
