% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_file_builders.R
\name{export_plotly}
\alias{export_plotly}
\title{Export Plotly to Portable HTML String}
\usage{
export_plotly(p, style = "height: 700px; width: 100\%; border: none;")
}
\arguments{
\item{p}{the plot to export}

\item{style}{the style attribute value for the plotly div}
}
\description{
This function exports plotly plots as stand alone iframes
}
\examples{
\dontrun{
export_plotly()
}
}
