% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding_helper.R
\name{field_info}
\alias{field_info}
\title{Field Info}
\usage{
field_info(fields, width = 160, return_packet = FALSE)
}
\value{
nothing but it prints info out
}
\description{
looks up info from the data dictionary for a study for a given field
}
\examples{
\dontrun{
field_info()
}
}
