% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_handler.R
\name{fields_from_similar_function_different_study}
\alias{fields_from_similar_function_different_study}
\title{Fields From Similar Function Different Study}
\usage{
fields_from_similar_function_different_study(
  name,
  preference = NULL,
  return_similar_name = FALSE,
  recursive = FALSE,
  injected_warning = NULL
)
}
\arguments{
\item{name}{the name of construct}

\item{preference}{a function name to use to extract preferred study}
}
\value{
fields character vector
}
\description{
returns the fields a similar construct uses
}
\examples{
\dontrun{
fields_from_similar_function_different_study()
}
}
