% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_file_builders.R
\name{figure}
\alias{figure}
\title{Wrap HTML snippet in a figure tag with an optional caption and number}
\usage{
figure(html, caption = NULL, caption_number = NULL)
}
\arguments{
\item{html}{A character string containing the HTML snippet to be wrapped.}

\item{caption}{An optional character string for the figure caption. If `NULL`,
no caption will be added.}

\item{caption_number}{An optional integer or string to prefix the caption 
with a number.}
}
\value{
The function does not return a value but prints the HTML code with 
the figure and caption tags.
}
\description{
This function takes an HTML snippet and wraps it in a `<figure>` tag. If a 
caption is provided, it will be added within a `<figcaption>` tag. If 
`caption_number` is provided, the caption will be prefixed with 
"Figure [caption_number]: ".
}
\examples{
\dontrun{
figure()
}
}
