% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymizer.R
\name{filter_redcap_data}
\alias{filter_redcap_data}
\title{Filter Redcap Data}
\usage{
filter_redcap_data(filter_column)
}
\arguments{
\item{filter_column}{name of column to filter by}
}
\value{
nothing
}
\description{
Helps address phi concerns when distributing redcap data by removing screened
}
\examples{
\dontrun{
Filter_redcap_data()
}
}
