% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_explorer.R
\name{find_all_compatible_functions}
\alias{find_all_compatible_functions}
\title{Find all compatible functions}
\usage{
find_all_compatible_functions(
  return_fields = FALSE,
  quiet = FALSE,
  recursive = TRUE,
  everything = FALSE,
  verify = FALSE
)
}
\arguments{
\item{return_fields}{defaults to FALSE and when TRUE returns the fields in the compatible functions rather than function names}

\item{quiet}{defaults to FALSE and when TRUE hides added data from being printed}

\item{recursive}{check to see if the function could work if we use another function}

\item{everything}{include legacy functions}

\item{verify}{if not set to recursive determine compatibility by running the function}
}
\value{
vector of characters that are the names of the compatible functions
}
\description{
returns vector of all the construct functions in this package that can be used with this study
}
\examples{
\dontrun{
find_all_compatible_functions()
}
}
