% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_explorer.R
\name{find_all_constructs}
\alias{find_all_constructs}
\title{Find all constructs}
\usage{
find_all_constructs(within_study = TRUE, auto_cache_ignore = FALSE)
}
\arguments{
\item{study}{if study is passed then functions will return names of constructs only for that study}
}
\value{
vector of characters that are the names of the constructs
}
\description{
returns vector of all the construct functions in this package
}
\examples{
\dontrun{
find_all_constructs()
}
}
