% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_explorer.R
\name{find_all_fields}
\alias{find_all_fields}
\title{Find all fields}
\usage{
find_all_fields()
}
\value{
vector of characters that are the names of the fields
}
\description{
returns vector of all the fields used by this study in this package
}
\examples{
\dontrun{
find_all_fields()
}
}
