% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_formatting.R
\name{format_count_percent}
\alias{format_count_percent}
\title{Format Count Percent}
\usage{
format_count_percent(numerator, denominator, decimals = 0)
}
\arguments{
\item{numerator}{single value}

\item{denominator}{single value}

\item{decimals}{number of decimals places, defaults to zero}
}
\value{
string
}
\description{
Returns the numerator and in parenthesis the calculated percentage of denominator
}
\examples{
\dontrun{
format_count_percent()
}
}
