% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_formatting.R
\name{format_mean_sd}
\alias{format_mean_sd}
\title{Format Mean Standand Deviation}
\usage{
format_mean_sd(values, decimals = 2)
}
\arguments{
\item{values}{insert values that you request mean and standard deviation for}

\item{decimals}{number of decimals places, defaults to zero}
}
\value{
string
}
\description{
Returns the mean and standard deviation for the requested vector(numbers)
}
\examples{
\dontrun{
format_mean_sd()
}
}
