% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptor.R
\name{function_to_codeblock}
\alias{function_to_codeblock}
\title{Function to Code Block}
\usage{
function_to_codeblock(
  name,
  data_name = "redcap_export",
  analytic_name = "analytic_dataset",
  codebook = NULL,
  width = 100,
  first = FALSE
)
}
\arguments{
\item{data_name}{name of full redcap dataframe}

\item{analytic_name}{name of analytic dataset table}

\item{codebook}{pre downloaded data dictionary}

\item{width}{width to truncate comments}

\item{names}{specific study function}
}
\value{
code block
}
\description{
makes a block of code
}
\examples{
\dontrun{
function_to_codeblock()
}
}
