% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptor.R
\name{function_to_script}
\alias{function_to_script}
\title{Functions to Script}
\usage{
function_to_script(
  names,
  data_name = "redcap_export",
  analytic_name = "analytic_dataset",
  width = 100,
  return_lines = TRUE,
  folder = NULL,
  display_order = NULL
)
}
\arguments{
\item{names}{specific study function}

\item{data_name}{name of full redcap dataframe}

\item{analytic_name}{name of analytic dataset table}

\item{width}{width to truncate comments}

\item{return_lines}{activates return of lines defaults to true}
}
\value{
code block
}
\description{
makes a block of code
}
\examples{
\dontrun{
function_to_codeblock()
}
}
