% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_writer.R
\name{generate_data_dictionary}
\alias{generate_data_dictionary}
\title{Generate Data Dictionary}
\usage{
generate_data_dictionary(df)
}
\value{
nothing
}
\description{
generates simple
}
\examples{
\dontrun{
generate_data_dictionary()
}
}
