% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_data_handler.R
\name{get_associated_data}
\alias{get_associated_data}
\title{Get Associated Data}
\usage{
get_associated_data(sheet_name, col_types = NULL)
}
\arguments{
\item{sheet_name}{the name of the Google Sheet}

\item{col_types}{set the col_types of the Google Sheet}
}
\value{
nothing
}
\description{
Downloads sheets from study folder on Google Drive by name
}
\examples{
\dontrun{
get_associated_data()
}
}
