% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_data_handler.R
\name{get_associated_data_folder}
\alias{get_associated_data_folder}
\title{Get Associated Data Folder}
\usage{
get_associated_data_folder(folder_name, col_types = NULL)
}
\arguments{
\item{folder_name}{the name of the Google Drive Folder}

\item{col_types}{et the col_types of every Google Sheet in specified folder}
}
\value{
nothing
}
\description{
Downloads all sheets from study folder on Google Drive by subdirectory name
}
\examples{
\dontrun{
get_associated_data_folder()
}
}
