% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_data_handler.R
\name{get_codebook}
\alias{get_codebook}
\title{Get Redcap Code Book}
\usage{
get_codebook()
}
\value{
dataframe of Redcap Code Book
}
\description{
Redcap keeps a code book mapping the fields from the full dataset 
to the forms and questions they come from. This function downloads that codebook
provided you have already called set_data(TODO_STUDY, TODO_REDCAPKEY)
}
\examples{
\dontrun{
get_codebook()
}
}
