% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals_manager.R
\name{get_columns}
\alias{get_columns}
\title{Get Columns}
\usage{
get_columns()
}
\value{
vector of column names
}
\description{
get all columns from data file, though they may not all be loaded at that time
}
\examples{
\dontrun{
get_columns()
}
}
