% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_handler.R
\name{get_construct}
\alias{get_construct}
\title{Get Construct}
\usage{
get_construct(name)
}
\arguments{
\item{name}{the name of a construct function such that paste(study, name, sep="_") is fx name}
}
\value{
a function
}
\description{
returns the construct function matching the name input and set study
}
\examples{
\dontrun{
get_construct()
}
}
