% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_handler.R
\name{get_construct_fields}
\alias{get_construct_fields}
\title{Get Construct Fields}
\usage{
get_construct_fields(name, recursive = TRUE)
}
\arguments{
\item{name}{the name of a construct function such that paste(study, name, sep="_") is fx name}

\item{recursive}{whether to recursively get the fields of the functions called by the construct}
}
\value{
fields character vector
}
\description{
returns the fields a construct function uses
}
\examples{
\dontrun{
get_construct_fields()
}
}
