% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_handler.R
\name{get_construct_function_name}
\alias{get_construct_function_name}
\title{Get Construct Function Name}
\usage{
get_construct_function_name(name)
}
\arguments{
\item{name}{the name of a construct function}
}
\value{
a character string
}
\description{
returns the name of the construct function matching the name input and set study
}
\examples{
\dontrun{
get_construct_function_name()
}
}
