% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_handler.R
\name{get_construct_output}
\alias{get_construct_output}
\title{Get Construct Output}
\usage{
get_construct_output(name)
}
\arguments{
\item{name}{the name of a construct function such that paste(study, name, sep="_") is fx name}
}
\value{
standard construct output
}
\description{
returns output of the construct function matching the name input and set study
}
\examples{
\dontrun{
get_construct_output()
}
}
