% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_handler.R
\name{get_construct_outputs}
\alias{get_construct_outputs}
\title{Get Construct Outputs}
\usage{
get_construct_outputs(names)
}
\arguments{
\item{names}{the vector of the names of the construct function such that paste(study, name, sep="_") is fx name}
}
\value{
standard construct output
}
\description{
returns joined output of the construct function matching the name input and set study
}
\examples{
\dontrun{
get_construct_outputs()
}
}
