% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_data_handler.R
\name{get_data}
\alias{get_data}
\title{Get Data}
\usage{
get_data(selected_columns)
}
\arguments{
\item{selected_columns}{this is a vector of characters that can use to specifing desired columns}
}
\value{
dataframe of some Redcap data
}
\description{
provided you have already called set_data(TODO_STUDY, TODO_REDCAPKEY),
but remember set_data should NEVER be called inside a constructor function,
than this function can be used inside of constructor functions by specifing the columns
you would like from the full dataset
}
\examples{
\dontrun{
get_data()
}
}
