% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_data_handler.R
\name{get_data_developer}
\alias{get_data_developer}
\title{Get Data Devleoper}
\usage{
get_data_developer()
}
\value{
dataframe of all Redcap data
}
\description{
provided you have already called set_data(TODO_STUDY, TODO_REDCAPKEY),
but remember set_data should NEVER be called inside a constructor function,
than this function can be used inside of constructor functions while they are in development,
but to improve performance after development get_data should be used with needed columns specified
}
\examples{
\dontrun{
get_data_developer()
}
}
