% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_handler.R
\name{get_dependencies}
\alias{get_dependencies}
\title{Get Dependencies}
\usage{
get_dependencies(name)
}
\value{
vector of dependencies
}
\description{
finds what constructs a construct function relies on
}
\examples{
\dontrun{
get_dependencies()
}
}
