% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_handler.R
\name{get_function_fields}
\alias{get_function_fields}
\title{Get Function Fields}
\usage{
get_function_fields(name, recursive = TRUE, injected_warning = NULL)
}
\arguments{
\item{name}{the name of a function}
}
\value{
fields character vector
}
\description{
returns the fields a function uses
}
\examples{
\dontrun{
get_function_fields()
}
}
