% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_tools.R
\name{if_needed_generate_example_data}
\alias{if_needed_generate_example_data}
\title{Generate dummy dataset}
\usage{
if_needed_generate_example_data(
  test_analytic,
  example_constructs = "",
  example_types = ""
)
}
\arguments{
\item{function_name}{name of the visualization function}

\item{key}{verification key}
}
\value{
summed tibble
}
\description{
Makes sure a related visualization has not changed
}
\examples{
\dontrun{
if_needed_generate_example_data()
}
}
