% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_file_builders.R
\name{install_default_files}
\alias{install_default_files}
\title{Install Default Files for R Markdown Support}
\usage{
install_default_files(target_path)
}
\arguments{
\item{target_path}{install path}
}
\value{
nothing
}
\description{
Installs all files needed to support R Markdown files instead of R scripts
}
\examples{
\dontrun{
install_default_files()
}
}
