% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymizer.R
\name{is_anonymizable}
\alias{is_anonymizable}
\title{is anonymizable}
\usage{
is_anonymizable(path)
}
\arguments{
\item{path}{path to script}
}
\value{
boolean
}
\description{
this helps you determine if the code created can really be used with anonymized data.
It checks for hardcoded dates and study_ids
}
\examples{
\dontrun{
is_anonymizable()
}
}
