% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding_helper.R
\name{kickstart_function}
\alias{kickstart_function}
\title{New Function Kickstarter}
\usage{
kickstart_function(
  function_name,
  field,
  construct = "screened",
  event = NA,
  instance = NA,
  instrument = NA
)
}
\arguments{
\item{function_name}{name of the function you want to build (sans the study name)}

\item{field}{name of a redcap field to start with}

\item{construct}{name of construct to start with}

\item{event}{event to filter in}

\item{instance}{event to filter in}

\item{instrument}{event to filter in}
}
\value{
string of function
}
\description{
you can pass it any vector and it will format it
}
\examples{
\dontrun{
kickstart_function()
}
}
