% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_writer.R
\name{legacy_generate_data_dictionary}
\alias{legacy_generate_data_dictionary}
\title{Legacy Generate Data Dictionary}
\usage{
legacy_generate_data_dictionary(df)
}
\value{
nothing
}
\description{
generates simple
}
\examples{
\dontrun{
legacy_generate_data_dictionary()
}
}
