% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_file_builders.R
\name{metrc_basic_report}
\alias{metrc_basic_report}
\title{METRC Basic Report}
\usage{
metrc_basic_report(html_table_list, file_name)
}
\arguments{
\item{html_table_list}{list of named html tables (set the name in the list to be the table name)}

\item{file_name}{place to store html file}
}
\value{
nothing
}
\description{
This function gets the payments for a study using the stored method
}
\examples{
\dontrun{
metrc_basic_report()
}
}
