% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_file_builders.R
\name{metrc_basic_report_numbers_labels}
\alias{metrc_basic_report_numbers_labels}
\title{METRC Basic Report Numbers and Labels}
\usage{
metrc_basic_report_numbers_labels(html_table_list, numbers, labels, file_name)
}
\arguments{
\item{html_table_list}{list of named html tables (set the name in the list to be the table name)}

\item{numbers}{vector of the numbers for the html tables (must match length of html_table_list)}

\item{labels}{vector of the labels for the html tables (ex. Table or Figure)}

\item{file_name}{place to store html file}
}
\value{
nothing
}
\description{
This function gets the payments for a study using the stored method
}
\examples{
\dontrun{
metrc_basic_report_numbers_labels()
}
}
