% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_writer.R
\name{order_functions}
\alias{order_functions}
\title{Order Functions}
\usage{
order_functions(names = NULL)
}
\value{
vector of function names ordered
}
\description{
orders the names of a study
}
\examples{
\dontrun{
order_functions()
}
}
