% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding_helper.R
\name{pretty_print_cols}
\alias{pretty_print_cols}
\title{Pretty prints a vector so it can be pasted into a vecort}
\usage{
pretty_print_cols(vec)
}
\arguments{
\item{vec}{any vector}
}
\value{
string of vector
}
\description{
you can pass it any vector and it will format it
}
\examples{
\dontrun{
pretty_print_cols()
}
}
