% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymizer.R
\name{randomize_tibble}
\alias{randomize_tibble}
\title{Randomize Tibble}
\usage{
randomize_tibble(data)
}
\arguments{
\item{data}{Tibble}
}
\value{
nothing
}
\description{
randomizes each column of tibble with replacement
}
\examples{
\dontrun{
randomize_tibble()
}
}
