% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_tools.R
\name{reorder_rows}
\alias{reorder_rows}
\title{Reorder Rows}
\usage{
reorder_rows(df, new_row_order, unspecified_to_bottom = TRUE)
}
\arguments{
\item{df}{working dataframe}

\item{new_row_order}{string vector of new order of things}

\item{unspecified_to_bottom}{whether values not specified will be on the top or bottom}
}
\value{
dataframe
}
\description{
sreorders the rows of a dataframe based on given vectors
}
\examples{
\dontrun{
reorder_rows()
}
}
