% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requirer.R
\name{require_automatic_cache}
\alias{require_automatic_cache}
\title{Require Automatic Cache}
\usage{
require_automatic_cache(force = FALSE)
}
\value{
nothing
}
\description{
updates the automatic cache
}
\examples{
\dontrun{
require_automatic_cache()
}
}
