% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requirer.R
\name{require_study}
\alias{require_study}
\title{Require Study}
\usage{
require_study()
}
\value{
nothing
}
\description{
throws an error if the study has not been set
}
\examples{
\dontrun{
require_study()
}
}
