% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding_helper.R
\name{search_fields}
\alias{search_fields}
\title{Search Fields Info}
\usage{
search_fields(terms, crf = NULL)
}
\value{
nothing but it prints info out
}
\description{
looks up info from the data dictionary for a study for a given search term
}
\examples{
\dontrun{
search_fields()
}
}
