% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_data_handler.R
\name{set_data}
\alias{set_data}
\title{Set Data}
\usage{
set_data(
  study,
  key,
  date = NULL,
  folder = NULL,
  server = FALSE,
  columns = NULL,
  auto_column_mode = "detect"
)
}
\arguments{
\item{study}{this is the name of the study recognized by the metrc-reports server}

\item{key}{this is the redcap api key for that study}

\item{date}{if you want to select a specific day the data was cached input it in 0000-00-00 format}

\item{folder}{if you want to store your cached data in a custom folder}

\item{server}{this tells the package that it is within a server environment and to look for local files as data soruces}

\item{columns}{this allows you to prespecify the columns you want to use}

\item{auto_column_mode}{this only is utilized if columns are not NULL and if auto_column_mode is "defined"
it selects the columns in relevant construct functions, if it is set to "detect" it will include any fields
from any functions that use columns present in the data, and if it is set to something else or "" only the columns
in the columns argument will be selected}
}
\value{
nothing
}
\description{
Redcap keeps a code book mapping the fields from the full dataset 
to the forms and questions they come from. This function downloads that codebook
provided you have already called set_data(TODO_STUDY, TODO_REDCAPKEY)
}
\examples{
\dontrun{
set_data()
}
}
