% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_data_handler.R
\name{set_data_custom}
\alias{set_data_custom}
\title{Set Data Custom}
\usage{
set_data_custom(data_bundle, study, folder, server = FALSE)
}
\arguments{
\item{study}{this is the name of the study recognized by the metrc-reports server}

\item{folder}{this locates the data}

\item{server}{this tells the package that it is within a server environment and to look for local files as data soruces}

\item{data}{this is and object with an existing tibble in it or and existing tibble}
}
\value{
nothing
}
\description{
This function loads data directly from an existing tibble
}
\examples{
\dontrun{
set_data_custom()
}
}
