% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_data_handler.R
\name{set_function_sheet}
\alias{set_function_sheet}
\title{Set Function Sheet}
\usage{
set_function_sheet(fdf, study)
}
\arguments{
\item{fdf}{this is and object with an existing tibble in it}

\item{study}{this is the name of the study recognized by the metrc-reports server}
}
\value{
nothing
}
\description{
This function loads the function sheet from a tibble
}
\examples{
\dontrun{
set_function_sheet()
}
}
