% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_explorer.R
\name{study_uses_legacy}
\alias{study_uses_legacy}
\title{Determine if a study is using the legacy matrix}
\usage{
study_uses_legacy()
}
\value{
boolean
}
\description{
returns TRUE if the study is using the legacy matrix
}
\examples{
\dontrun{
study_uses_legacy()
}
}
