% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_tools.R
\name{summate_levels}
\alias{summate_levels}
\title{Summate Levels}
\usage{
summate_levels(df, group_colnames, level_colname, levels)
}
\arguments{
\item{df}{working dataframe}

\item{group_colnames}{groups to have in output}

\item{level_colname}{column of values to keep together}

\item{levels}{values to combine together}
}
\value{
dataframe
}
\description{
Gives the sum of all the columns in a table
}
\examples{
\dontrun{
summate_levels()
}
}
