% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_file_builders.R
\name{table}
\alias{table}
\title{Add a caption to an HTML table snippet with an optional number}
\usage{
table(html, caption = NULL, caption_number = NULL)
}
\arguments{
\item{html}{A character string containing the HTML table snippet.}

\item{caption}{An optional character string for the table caption. If `NULL`,
no caption will be added.}

\item{caption_number}{An optional integer or string to prefix the caption 
with a number.}
}
\value{
The function does not return a value but prints the modified HTML 
code with the added caption.
}
\description{
This function takes an HTML snippet containing a table and adds a `<caption>` 
tag to it. If a caption is provided, it will be inserted after the opening 
`<table>` tag. If `caption_number` is provided, the caption will be prefixed 
with "Table [caption_number]: ". If no caption is provided, the function 
outputs the snippet as is.
}
\examples{
# Example usage:
# Create an HTML table snippet
html_table <- "<table><tr><td>Example</td></tr></table>"

# Call table() to add a caption to the table with a number
table(html_table, "This is a table caption", 1)
\dontrun{
table()
}
}
