% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_tester.R
\name{test_code}
\alias{test_code}
\title{Test Construct Function Code}
\usage{
test_code(name)
}
\arguments{
\item{name}{the name of the construct excluding the starting study name}
}
\value{
character string of issues
}
\description{
checks whether the code of a construct function is compliant with standard output requirments
}
\examples{
\dontrun{
test_code()
}
}
