% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_tester.R
\name{test_documentation}
\alias{test_documentation}
\title{Test Construct Function Documentation}
\usage{
test_documentation(name)
}
\arguments{
\item{name}{the name of the construct excluding the starting study name}
}
\value{
character string of issues
}
\description{
checks whether the documentation of a construct function is compliant with standard output requirments
}
\examples{
\dontrun{
test_documentation()
}
}
