% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_tester.R
\name{test_output}
\alias{test_output}
\title{Test Construct Function Output}
\usage{
test_output(name, ids = NULL)
}
\arguments{
\item{name}{the name of the construct excluding the starting study name}

\item{ids}{the full list of study ids, can be used to increase performance when this is already known}
}
\value{
character string of issues
}
\description{
checks whether the output of a construct function is compliant with standard output requirments
}
\examples{
\dontrun{
test_output()
}
}
