% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_tester.R
\name{test_study}
\alias{test_study}
\title{Test Study Construct Functions}
\usage{
test_study()
}
\value{
table of test results
}
\description{
goes through each construct and tests documentation, code, and output for compliance with data standards
}
\examples{
\dontrun{
test_study()
}
}
