% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_filer.R
\name{unzip_construct}
\alias{unzip_construct}
\title{Unzip a Construct}
\usage{
unzip_construct(name)
}
\arguments{
\item{name}{name of the construct}
}
\value{
tibble unzipped
}
\description{
Helps unzip a zipped construct using info in documentation
}
\examples{
\dontrun{
unzip_construct(name)
}
}
