% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_sheet_utilities.R
\name{update_function_sheet_for_study}
\alias{update_function_sheet_for_study}
\title{Update Function Sheet Study Status}
\usage{
update_function_sheet_for_study(retries = 5)
}
\arguments{
\item{retries}{The number of times to retry updating the function sheet if it is locked by another process}
}
\value{
This function does not return anything
}
\description{
This function updates the study status of the function sheet for the current study
}
\examples{
\dontrun{
update_function_sheet_for_study()
}
}
