% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_sheet_utilities.R
\name{update_function_sheet_metadata}
\alias{update_function_sheet_metadata}
\title{Update Function Sheet Metadata}
\usage{
update_function_sheet_metadata(retries = 5)
}
\arguments{
\item{retries}{number of times to retry the update if it fails}
}
\value{
nothing
}
\description{
updates the metadata for the function sheet by pulling the latest version from google sheets
then checking to see if the "Function Name" is in the package and if not setting the "Status" to "Error" and the
"System Message" to "Function not found in package". Then upload these changes to google sheets.
}
\examples{
\dontrun{
update_function_sheet_metadata()
}
}
