% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_filer.R
\name{write_all_zipped_constructs}
\alias{write_all_zipped_constructs}
\title{Write all zipped constructs}
\usage{
write_all_zipped_constructs(dir_path, name = NULL, constructs = NULL)
}
\arguments{
\item{dir_path}{the directory to save the files to}

\item{name}{name of the constructs to join, defaults to null and won't be included}

\item{constructs}{list of constructs to consider unzipping}
}
\value{
nothing
}
\description{
Helps unzip a zipped construct using info in documentation
}
\examples{
\dontrun{
write_all_zipped_constructs(name)
}
}
