% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_data_handler.R
\name{write_data_to_excel}
\alias{write_data_to_excel}
\title{write data to excel}
\usage{
write_data_to_excel(dir_path)
}
\arguments{
\item{dir_path}{the directory to file to convert to xlsx}
}
\value{
nothing
}
\description{
converts csv at path to exlsx at same path
}
\examples{
\dontrun{
write_data_to_excel(name)
}
}
