% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_writer.R
\name{write_study}
\alias{write_study}
\title{Write Study}
\usage{
write_study(
  study,
  key,
  column_names,
  data_path,
  error_path,
  date = NULL,
  folder = NULL,
  server = FALSE,
  random = TRUE,
  dictionary = TRUE,
  no_treat = TRUE,
  anonymize = TRUE,
  anon_allow = FALSE,
  filter_column = NULL,
  long_file_additions = NULL
)
}
\arguments{
\item{study}{the study you want to run}

\item{key}{the Redcap api key to that study}

\item{column_names}{the column names you want in the dataset if you pass null all will be used}

\item{data_path}{path to write analytic set to}

\item{error_path}{path to write errors to}

\item{date}{set date of dataset requested}

\item{folder}{set the folder the raw data is located in defaults to STUDYNAME-cached_data}

\item{server}{sets cached data selector to server mode}

\item{random}{scrambles the data frame - this is not going to return accurate results but can be used to check code}

\item{dictionary}{includes a generated data dictionary}

\item{no_treat}{excludes treat field}

\item{anonymize}{creates random study_ids and date shifts all data}

\item{anon_allow}{allow anonymization despite issues}

\item{long_file_additions}{this is a string with comma separations that is used to add columns to long files of constructs}
}
\value{
nothing
}
\description{
returns vector of all inappropriately named construct functions in this package
}
\examples{
\dontrun{
write_study()
}
}
