% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_tools.R
\name{if_needed_generate_example_data}
\alias{if_needed_generate_example_data}
\title{Generate dummy dataset}
\usage{
if_needed_generate_example_data(
  test_analytic,
  example_constructs = "",
  example_types = ""
)
}
\value{
dummy analytic dataset
}
\description{
Generates random data to test visualizations. Currently accepted data types are:
Category,
Boolean,
Date,
Number,
Number1dgt,
FacilityCode (15 Codes, AAA - AAO),
Character (random lorem string, 1-3 "sentences"),
Form,
Period,
FollowupStatus,
FollowupPeriod,
NamedCategory,
TreatmentArm

Adding a '-N' to the end of a type will replicate values, separated with a comma.
'-NS' will separate replicated values with a semicolon.
The '-U(n)' tag allows you to specify the size of the category or range of numbers 
you'd like to generate data for. Replace (n) with the integer size of the category 
or number range. EX: \'Number-U4\'
Long Files are supported and are formed like
"('rowsep', 'colsep')Type1|Type2|Type3" etc.
NamedCategory type is formed like
"NamedCategory[\'Value 1\' \'Value 2\']" etc.
}
